# patient-ms/src/main/resources/static/openapi/patient-ms-patient-api.yaml
openapi: 3.0.3
info:
  title: Patient MS Patient API
  description: Patient MS Patient API
  version: 1.0.0

servers:
  - url: "/"
    description: "Current"
  - url: "http://localhost:8080"
    description: "localhost"

tags:
  - name: patients
    description: Everything for patient MS about patient

paths:
  /patient:
    get:
      tags:
        - patient
      summary: get all patients
      operationId: getAllPatients
      responses:
        '200':
          description: list of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags:
        - patient
      summary: create a patient
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreateDto'
      responses:
        '201':
          description: patient created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDto'

        '400':
          description: Invalid input

        '401':
          description: Unauthorized

        '403':
          description: Forbidden

  /patient/{id}:
    get:
      tags:
        - patient
      summary: get a patient by id
      operationId: getPatientById
      parameters:
        - in: path
          name: id
          required: true
          description: the id of the patient being queried
          schema:
              type: integer
              format: int64
              minimum: 1
      responses:
        '200':
          description: patient found for the given id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDto'
        '400':
          description: Invalid input
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Not found - no patient found for the given id
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    PatientDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: "The unique identifier of the patient"
          example: 1
        firstName:
          type: string
          description: "the first name of the patient"
          example: "John"
        lastName:
          type: string
          description: "the last name of the patient"
          example: "Doe"
        birthDate:
          type: string
          format: date
          description: "the birth date of the patient in format yyyy-MM-dd"
          example: "1980-01-01"
        gender:
          type: string
          enum: ["M", "F"]
          description: "the gender of the patient"
          example: "M"
        address:
          type: string
          description: "the address of the patient"
          example: "1 High Street"
        phoneNumber:
          type: string
          description: "the phone number of the patient"
          example: "0123456789"
    PatientCreateDto:
      type: object
      properties:
        firstName:
          type: string
          description: "the first name of the patient"
          x-field-extra-annotation: |
            @NotBlank(message = "First name is required")
            @Size(max = 100, message = "First name must not exceed 100 characters")
          example: "John"
        lastName:
          type: string
          description: "the last name of the patient"
          x-field-extra-annotation: |
            @NotBlank(message = "Last name is required")
            @Size(max = 100, message = "Last name must not exceed 100 characters")
          example: "Doe"
        birthDate:
          type: string
          format: date
          description: "the birth date of the patient in format yyyy-MM-dd"
          x-field-extra-annotation: |
            @NotNull(message = "Birthdate is mandatory")
            @Past(message = "Birthdate must be a past date")
          example: "1980-01-01"
        gender:
          type: string
          enum: ["M", "F"]
          description: the gender of the patient
          x-field-extra-annotation: |
            @NotNull(message = "Gender is mandatory")
          example: "M"
        address:
          type: string
          description: "the address of the patient"
          x-field-extra-annotation: |
            @Size(min= 5 ,max = 50, message = "Address must not exceed 50 characters")
            @Pattern(regexp = "^\\d+\\s+[A-Za-z0-9\\s]+$", message = "Address must start with a number followed by the street name")
          example: "1 High Street"
        phoneNumber:
          type: string
          description: "the phone number of the patient"
          x-field-extra-annotation: |
            @Pattern(regexp = "^(\\d{10}|\\d{3}-\\d{3}-\\d{4})$", message = "Phone number must be exactly 10 digits or in the format xxx-xxx-xxxx")
          example: "0123456789"